Some cleanup on types
