+++ src/components/Work/Work.tsx
@@ -4,8 +4,7 @@ import * as DS from "@nypl/design-system-react-components";
 import { joinArrayOfElements } from "~/src/util/Util";
 import { EditionCard } from "~/src/components/EditionCard/EditionCard";
 import WorkDetailDefinitionList from "~/src/components/WorkDetailDefinitionList/WorkDetailDefinitionList";
-import { WorkResult } from "~/src/types/WorkQuery";
-import { ApiWork } from "~/src/types/DataModel";
+import { ApiWork, WorkResult } from "~/src/types/WorkQuery";
 import { breadcrumbTitles } from "~/src/constants/labels";
 import EditionCardUtils from "~/src/util/EditionCardUtils";
 import SearchHeader from "../SearchHeader/SearchHeader";
