+++ src/types/WorkQuery.ts
@@ -1,4 +1,5 @@
-import { ApiWork } from "./DataModel";
+import { Agent, Language, Subject } from "./DataModel";
+import { ApiEdition } from "./EditionQuery";
 
 export type WorkQuery = {
   identifier: string;
@@ -11,4 +12,22 @@ export type WorkResult = {
   timestamp?: string;
   responseType?: string;
   data?: ApiWork;
+};
+
+export type ApiWork = {
+  alt_titles?: string[];
+  authors?: Agent[];
+  contributors?: string[];
+  dates?: Date[];
+  editions?: ApiEdition[];
+  edition_count?: number;
+  languages?: Language[];
+  measurements?: string[];
+  medium?: string;
+  series?: string;
+  series_position?: string;
+  sub_title?: string;
+  subjects?: Subject[];
+  title?: string;
+  uuid?: string;
 };
